refactor: Changed structure folder  componetents - Changed structure
