+++ src/components/Authorization/AuthorizationControls/RegisterControl.tsx
@@ -1,11 +1,11 @@
 import { Form, Formik } from 'formik';
-import { InputFormik } from '../../../Input/FormikInput';
-import { Button } from '../../../Button/Button';
-import { registerFormFieldsDefault } from '../../../utils/auth';
-import { AuthControlProps } from '../AuthModalContent';
-import { PASS_MAX, INPUT_LENGTH } from '../../../../constants/index';
-import classes from '../AuthModal.module.scss';
-import { validationRegister } from '../../../../schemas/auth-validator';
+import { InputFormik } from '../../Common/Inputs/FormikInput';
+import { Button } from '../../Common/Buttons/Button';
+import { registerFormFieldsDefault } from '../../utils/auth';
+import { AuthControlProps } from '../AuthorizationModalContent';
+import { PASS_MAX, INPUT_LENGTH } from '../../../constants/index';
+import classes from '../AuthorizationModal.module.scss';
+import { validationRegister } from '../../../schemas/auth-validator';
 import { LuSquare, LuChevronDownSquare } from 'react-icons/lu';
 import { useState } from 'react';
