+++ src/components/Common/Logo/Logo.tsx
@@ -1,7 +1,7 @@
 import { NavLink } from 'react-router-dom';
 import classes from './Logo.module.scss';
-import logo_footer from '../../assets/icons/logo.svg';
-import logo_header from '../../assets/icons/logo_header.svg';
+import logo_footer from '../../../assets/icons/logo.svg';
+import logo_header from '../../../assets/icons/logo_header.svg';
 
 export const Logo: React.FC<{ type?: 'header' | 'footer' }> = ({ type = 'header' }) => {
   const logo = type === 'header' ? logo_header : logo_footer;
