+++ src/components/Common/Slider/Slider.module.scss
