+++ src/components/Home/AdvantagesList/AdvantageItem/AdvantageItem.module.scss
