+++ src/components/Home/AdvantagesList/AdvantageItem/AdvantageItem.tsx
