+++ src/components/Home/AdvantagesList/AdvantagesList.module.scss
