+++ src/components/Home/AdvantagesList/AdvantagesList.tsx
