+++ src/components/Home/Hero/Hero.module.scss
@@ -11,7 +11,6 @@
   background-size: cover;
 
   &_content {
-    width: 100%;
     height: 100%;
     margin: 0 auto;
     display: flex;
