+++ src/components/Home/Presentation/PresentationItem/PresentationItem.module.scss
