+++ src/pages/NotFound/NotFound.tsx
@@ -1,0 +1,9 @@
+const NotFound = () => {
+  return (
+    <>
+      <h1>404</h1>
+    </>
+  );
+};
+
+export default NotFound;
