--- src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,9 +1,0 @@
-const NotFoundPage = () => {
-  return (
-    <>
-      <h1>404</h1>
-    </>
-  );
-};
-
-export default NotFoundPage;
